+++ src/Notifications.h
@@ -7,6 +7,7 @@ extern Kind kNotifCursorPos;
 extern Kind kNotifActionResponse;
 extern Kind kNotifPageInfo;
 extern Kind kNotifAdHoc;
+extern Kind kNotifThemeRelayout;
 
 using NotificationWndRemoved = Func1<NotificationWnd*>;